pyproject.toml
[tool.poetry]
name = "calendar"
version = "0.1.0"
description = ""
authors = ["Hông-Lan Botterman <honglan.botterman@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^11.0.0"
qrcode = "^8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
